# [performance] Rework home timeline query to use cache more