# Description

> If this is a code change, please include a summary of what you've coded, and link to the issue(s) it closes/implements.
>
> If this is a documentation change, please briefly describe what you've changed and why.

This pr further optimizes our home timeline db function to use the cache a bit more, and produce a simpler query plan overall.

It does this by passing an IN statement to the database with a list of all the account IDs that the requesting account follows.

While this does produce some pretty preposterous looking queries when an account has 100+ followers, it is still much faster than the subquery we were using previously (at least on SQLite).

Previous query plan:

![Complex multi-line query plan](https://github.com/superseriousbusiness/gotosocial/assets/31960611/51ea9f05-2ae2-4acf-a1e5-11c2d17c0716)

Previous rough measurements of total time spent inside HomeTimeline:

```
42.03077ms
4.483957ms
1.442441078s
1.438765929s
1.32123294s
1.353804379s
1.151259976s
1.036415224s
1.298738296s
1.268815314s
1.130251552s
1.213766837s
1.612718543s
1.477530961s
6.23008ms
7.076481ms
```

Query plan with this change:

![Simpler query plan](https://github.com/superseriousbusiness/gotosocial/assets/31960611/c77ba664-de0c-4353-a0a1-5f8920076ceb)

Rough measurements of total time spent inside HomeTimeline function with this change:

```
117.130545ms
6.138885ms
30.566447ms
4.003555ms
75.938819ms
47.130639ms
14.810726ms
10.806687ms
7.098044ms
6.881918ms
9.327144ms
8.117014ms
91.331842ms
32.523115ms
7.262103ms
8.745176ms
89.383523ms
37.40782ms
16.049556ms
12.582155ms
42.878887ms
21.919784ms
9.897436ms
```

So, it's something like 50-100x faster consistently!

Closes https://github.com/superseriousbusiness/gotosocial/issues/2147

Note, not sure yet how well this will scale if people start having 500, 1000, or more followers. Probably not that well but we'll see.

We have an alternative implementation available [here](https://github.com/superseriousbusiness/gotosocial/tree/home_timeline_query_fiddling_alternative) which uses a temporary table to make selects. It also works out faster than the previous method, but not as fast as this one. Measurements for that alternative implementation:

```
40.711501ms
6.887407ms
54.195054ms
7.569312ms
37.996598ms
81.833877ms
43.759319ms
95.888541ms
51.250977ms
314.097388ms
104.913237ms
57.236474ms
38.230601ms
204.390611ms
66.997762ms
47.914982ms
122.991578ms
47.148936ms
37.858711ms
36.143723ms
44.443219ms
95.345832ms
45.044735ms
86.031615ms
41.128506ms
71.316812ms
29.760012ms
```

## Checklist

Please put an x inside each checkbox to indicate that you've read and followed it: `[ ]` -> `[x]`

If this is a documentation change, only the first checkbox must be filled (you can delete the others if you want).

- [x] I/we have read the [GoToSocial contribution guidelines](https://github.com/superseriousbusiness/gotosocial/blob/main/CONTRIBUTING.md).
- [x] I/we have discussed the proposed changes already, either in an issue on the repository, or in the Matrix chat.
- [x] I/we have performed a self-review of added code.
- [x] I/we have written code that is legible and maintainable by others.
- [x] I/we have commented the added code, particularly in hard-to-understand areas.
- [x] I/we have made any necessary changes to documentation.
- [x] I/we have added tests that cover new code.
- [x] I/we have run tests and they pass locally with the changes.
- [x] I/we have run `go fmt ./...` and `golangci-lint run`.